+++ src/libraries/sprite.lua
@@ -9,6 +9,8 @@ local RobloxMaxImageSize = 1024
 --[=[
 	Animates the given sprite to play like a GIF.
 
+	@yields
+
 	@param image ImageLabel -- The image label that the sprite should be animated on
 	@param imageSize Vector2 -- The size of the image
 	@param frames Vector2 -- The amount of frames on both the X and Y axis